Chore - merge conflicst to update validateAndLogProjectDetails function
