+++ packages/core/src/codewhisperer/service/transformByQ/transformApiHandler.ts
@@ -24,13 +24,14 @@ import AdmZip from 'adm-zip'
 import globals from '../../../shared/extensionGlobals'
 import { CredentialSourceId, telemetry } from '../../../shared/telemetry/telemetry'
 import { codeTransformTelemetryState } from '../../../amazonqGumby/telemetry/codeTransformTelemetryState'
-import { calculateTotalLatency } from '../../../amazonqGumby/telemetry/codeTransformTelemetry'
+import { calculateTotalLatency, JDKToTelemetryValue } from '../../../amazonqGumby/telemetry/codeTransformTelemetry'
 import { MetadataResult } from '../../../shared/telemetry/telemetryClient'
 import request from '../../../common/request'
 import { projectSizeTooLargeMessage } from '../../../amazonqGumby/chat/controller/messenger/stringConstants'
 import { ZipExceedsSizeLimitError } from '../../../amazonqGumby/errors'
 import { writeLogs } from './transformFileHandler'
 import { AuthUtil } from '../../util/authUtil'
+import { getOpenProjects } from './transformProjectValidationHandler'
 
 export function getSha256(buffer: Buffer) {
     const hasher = crypto.createHash('sha256')
@@ -46,31 +47,6 @@ export async function getAuthType() {
         authType = 'awsId'
     }
     return authType
-}
-
-// log project details silently
-export async function validateAndLogProjectDetails() {
-    let telemetryJavaVersion = JDKToTelemetryValue(JDKVersion.UNSUPPORTED) as CodeTransformJavaSourceVersionsAllowed
-    let errorCode = undefined
-    try {
-        const openProjects = await getOpenProjects()
-        const validProjects = await validateOpenProjects(openProjects, true)
-        if (validProjects.length > 0) {
-            // validProjects[0].JDKVersion will be undefined if javap errors out or no .class files found, so call it UNSUPPORTED
-            const javaVersion = validProjects[0].JDKVersion ?? JDKVersion.UNSUPPORTED
-            telemetryJavaVersion = JDKToTelemetryValue(javaVersion) as CodeTransformJavaSourceVersionsAllowed
-        }
-    } catch (err: any) {
-        errorCode = err.code
-    } finally {
-        telemetry.codeTransform_projectDetails.emit({
-            passive: true,
-            codeTransformSessionId: codeTransformTelemetryState.getSessionId(),
-            codeTransformLocalJavaVersion: telemetryJavaVersion,
-            result: errorCode ? MetadataResult.Fail : MetadataResult.Pass,
-            reason: errorCode,
-        })
-    }
 }
 
 export function throwIfCancelled() {
